+++ ChangeLog.md
@@ -8,6 +8,7 @@
 * Add support for 32-bit pixel data (Anders Gustafsson, Cureos AB)
 * Better handling of default item in Get<> method (Anders Gustafsson, Cureos AB)
 * Support opening DicomFile from Stream (Anders Gustafsson, Cureos AB)
+* Add support SIGMOID VOI LUT function
 
 #### v1.0.21 (09/14/2012)
 * Add connection close event and socket error handlers to DicomService